consensus: Use ReadySetError for errors, not anyhow
Switch from using anyhow for the error types throughout the consensus
module to using ReadySetError. Since we want to actually be able
to *handle* these errors (specifically, react to failure to deserialize
controller state by starting fresh with an empty state) we need an enum
we can actually pattern match on rather than just the "string in a box"
that anyhow provides us.

Change-Id: I4962a5734d245f61fcf9fbca03df84b957f7ae9a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4353
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>