server: Wipe controller state if deserialization fails
If we fail to deserialize the state when loading the controller state
after a successful leader election, just log a warning and replace it
with an empty controller state. This drops all user-created caches,
which is non-ideal, but is at least better than the *current* status
quo, where the controller state would have to be wiped from the
authority manually (still dropping user-created caches, but requiring
manual operational intervention!).

This is relatively tricky to test automatically, but I've tested it
manually by introducing a backwards-incompatible change to the
deserialization of Config (a new field) and things seem to work well -
plus, since we aren't deleting rocksdb database files, we don't
resnapshot!

An interesting corollary of that last bit is that if a table gets
dropped in the upstream database while readyset isn't running, and then
we upgrade to a new version that requires dropping controller state, the
rocksdb database for that table will stick around forever - there's
basically no way around this right now, but it's an interesting thing to
keep an eye out for going forwards.

Release-Note-Core: When upgrading to a new version of ReadySet with an
  incompatible format for user configuration, ReadySet will now
  automatically wipe out the cluster state, where previously manual
  intervention was required. Note that this does mean that upgrading to
  a new version of ReadySet can currently result in losing `CREATE
  CACHE` statements.
Fixes: ENG-2561
Change-Id: Ic124e2e7f39019c0a3bbf5e9eee48818e545ed6a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4355
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>