readyset-server: fix flaky test failure
We've seen issues with tests that start a ReadySet server, shut it down,
and start a second one later in the same test: every once in a while,
the `Handle::backend_ready` hangs waiting for the second server to start
up. It seems the cause of this has to do with background tasks spawned
by the controller responding to  the shutdown signal before the controller
does. This can make it look like there was an error communicating with a
background task when, really, the task was just gracefully shut down.

This change adds checks to see if a shutdown signal is the cause of a
communication error between tasks. It also adds shutdown signal checking
in a few more background tasks for which shutdown wasn't being handled.

Change-Id: I7d5389de5d59f804cbc8d7942fcc32fccc73aacb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4362
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>