system-benchmarks: Allow comparing against upstream db
Add a new flag to system-benchmarks to also run benchmarks directly
against the upstream db, to compare our latency when caching or proxying
queries to that of the upstream.

Also includes some miscellaneous cleanup that didn't feel worth breaking
into their own commits given this is a developer tool, including some
cleanup around Results and a fixed typo in a constant name.

Change-Id: I4c36465164ae6533d4a887bd3be8c7633aa4e51d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4364
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>