readyset-psql: Avoid extra allocations in upstream execute
Refactor the upstream `execute` method to avoid allocating two
intermediate vectors, instead just allocating *one* vector for the
results. This sadly has no noticeable effect on the actual runtime or
throughput of the query proxying benchmark I'm running, but is such an
obvious optimization that I'm committing it anyway as a hopefully
incremental improvement.

Change-Id: Ia60c94a1436ce909d7a81fb555bc796a955dbbab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4367
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI