treewide: Remove all instances of #[doc(hidden)]
These made a little bit more sense when the readyset Rust interface was
a stable, public library that was intended to be used as rust
applications, but now that we're using Rustdoc more as developer
documentation it doesn't really make sense to be hiding any
documentation, anywhere.

This also removes the `#[deny(missing_docs)]` annotations on
readyset-client and readyset-server, since now there's a lot of
undocumented identifiers that I don't really want to make up docs for
just for this commit - having docs for everything is a noble goal but
probably unreasonable for a production application codebase like this
one.

Change-Id: I2408b72c6a67135b75089f572201f9bf50e25dd8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4380
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>