Add cleanup mode for persistent state
Adds a new cleanup flag which when passed will clean up the assets on
disk (rocksdb and auth) for the provided deployment. This also refactors
where we store these files. We now place them in a dedicated subfolder
for the deployment. This allows us to easily clean up deployment assets
by simply removing the deployment subfolder.

Change-Id: Ia79d255dea4f83ce8bdd915d2a448b82d23755e6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4385
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>