build: pin docker image versions
This change pins the docker image versions for the containers we run in
our CI pipelines. To do this, I had to:

1. Update the buildkite pipeline that mirrors images to ECR to pull
   specific versions of the images we care about instead of always
   pulling the image with the `latest` tag; and
2. Update the docker compose files/Dockerfiles to reflect these pinned versions.

Pinning these image versions is preferable because it ensures that we
control the parameters of our test environments, which makes it easier
to reproduce, diagnose, and fix errors and bugs.

Change-Id: Ifa7720da825160b29544f9ee72200de9e3c8a78e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4387
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Ian Meyer <ian@readyset.io>
Tested-by: Buildkite CI