readyset-psql: Stream results from upstream
Rather than accumulating results from upstream in a vector before even
*beginning* to send results downstream, return the stream of results
from upstream queries all the way down to psql-srv, sending results as
soon as we get them from the upstream db. In a very simple proxy app
that only uses psql-srv (eg not involving readyset-psql at all) this
increases QPS with query_slammer from around 200 to around 7000 - we
don't get *any* increase when running with readyset-psql so there's
likely some additional blocking going on there, but this is definitely a
step in the right direction.

Fixes: ENG-1754
Change-Id: Ib66cefa98a01ab76578e4d18f9f402ad5738c22e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4370
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>