ddl_vertical: Make sure ReadySet reads don't go to fallback
Without explicitly creating the cache and using the ALWAYS keyword,
errors can cause the postcondition query to ReadySet to fall back to
upstream, which prevents us from accurately testing whether the data in
Postgres matches the data in ReadySet.

Explicitly creating the cache usually succeeds right away, but seems to
require a number of retries when a CREATE follows an ALTER TABLE, which
I assume has something to do with the partial resnapshot needed after
the ALTER TABLE is executed.

Change-Id: I84225bef5848865645af78540d49d717b575fd2d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4399
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>