ddl_vertical: Fix precondition bug for WriteRow
Without this extra precondition, it's possible for WriteRow operations
to trigger a false positive test failure during shrinking due to the
table schema changing when an ALTER COLUMN step is removed.

Might be nice to try to add a "fix operation" callback to attempt to
regenerate the row values if the precondition fails for a given step,
but that's left as a project for a future commit.

Change-Id: I30d797d778bc7e6871af2e610f531d83d69cf621
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4440
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>