query_graph: Don't bail on unexpected filters
If we encounter an unexpected filter pattern when classifying predicates
for a query graph, it's probably just constant-valued (eg, we've seen
this get hit on `WHERE x = ? AND 1=1`, for example). Rather than
bailing (with a confusing error message!) we can just add the predicate
to the set of global predicates for the query - if it *does* in fact
contain placeholders, we'll just fail during lowering, but if it's
constant valued then we can turn it into a filter node and successfully
execute the query.

Fixes: ENG-2510
Change-Id: Ie822886fcc0588ba74fe9ab46a7c3681188a9918
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4424
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>