replicators: Properly replicate ALTER TABLE RENAME COLUMN
For the `RENAME COLUMN` DDL operation, the `object_type` column returned
by `pg_event_trigger_ddl_commands` is "table column", not "table", but
we were incorrectly only filtering for DDL events with "table". This
fixes that, covering it with both a test in ddl_replication.rs itself
and an integration test covering that we successfully replicate that
ALTER TABLE event.

Fixes: ENG-2627
Change-Id: I17cfa95a8f8077adaed167c8ab10eabeea5cf220
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4453
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>