ddl_vertical: Remove names from list of deleted tables when needed
This is unlikely to be an issue very often since we use random names for
tables and views, but I'm thinking about trying to restrict things to
increase the odds of reusing table/view names at times, so this might
become more relevant soon. But anyway, if we happened to create a table,
delete it, and then generate another table or view with the same name,
it could cause false positive test failures. This commit fixes that
problem.

Change-Id: I3d3453cffb7a4e154e3f8732e27e8a04677bff75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4459
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>