persistent-state: Normalize decimal values in keys
Decimal values that have the same value but different scales, eg "5.0"
and "5", compare and hash as equal correctly, but don't *serialize* to
the same value (because, obviously, the scale is part of the value).
This means that if we serialize a key with a Numeric value to RocksDB,
we aren't then able to do point lookups to that key with Numeric values
of a different scale.

Fortunately, Decimal exposes a `normalize()` method, which is what it
uses for its own Hash method, which converts the decimal to a
representation where two semantically-equal values will actually be
representationally-equal (by eg trimming off trailing zeros). We do
still need to preserve the non-normalized representation to
actually *return* (because the scale is observable), so this commit
modifies persistent state to call that normalize method on Numeric
values, but *only for keys*, not for the actual values in rows.

Fixes: ENG-2594
Change-Id: I0ed061d5c8b9c85b645293791d438976277e9941
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4463
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>