replicators: Improve logging of log positions
Improve the logging within replicators of log positions in a couple
places, by expanding out the logging of the fields of `system` so we get
the Display impl for PostgresPosition, and by loading and logging the
`confirmed_flush_lsn` of the replication slot we're using before we
start replicating. This improves debuggability of replication issues
having to do with log offsets.

Change-Id: I18c930dabcb646e76af59e2ffe7dd62405962600
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4467
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>