replicators: Resnapshot all tables if we don't have an offset
There are a couple of reasons that `ReplicationOffsets::max_offset`
might return `None`:

1. This is a fresh deployment with no tables, or no data for any tables
2. *some* tables, but not all, are missing a replication offset for some
   reason (eg one of them hasn't gotten any writes!)

In *both* cases, we drop and recreate the replication slot on startup of
the replicator - but in the second case, this means that any writes to
the tables that *do* have replication offsets will never be able to be
replicated!

To fix that issue in the safest possible way, this commit just makes it
so that if we don't have a consistent replication offset when we're
resnapshotting, we drop and resnapshot *all* tables, rather than just
the ones that don't have a replication offset. This is crappy and slow,
but will *definitely* get the data to a consistent point, and we can
patch over other issues that get us to this point as an optimization
later.

Fixes: ENG-2655
Change-Id: I71d2076d8c725c8346ced531fb4419291eeaed31
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4468
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>