Clustertests: Remove first_query condition in query_until_expected_inner
e47aba4 introduced a condition to query_until_expected_inner() that
causes us to believe that we always use the correct source on the first
iteration of the loop. This CL assumes that the behavior was introduced
in error and removes the condition.

With the above change, a few tests required the addition of CREATE CACHE
statements to succeed `query_until_expected_from_noria` calls. This CL
also adds a query_until_ok() function to facilitate running CREATE CACHE
statements when the controller may not be ready.

Change-Id: I82d4cacd555d76455f0e39e6bb8de02f422403cd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4471
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>