client-test-helpers: Allow configuring db name
Rather than *always* hardcoding the DB name to use when running
TestBuilder with an upstream, add methods to optionally configure the db
name. Everything that was previously using the hardcoded db name still
is here, but this paves the way to being able to run certain (at least
postgres-backed) tests in parallel by having them point at a different
database.

Change-Id: I1aaf0a1db8ecf17cbe894d8a11be72144bb1e658
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4481
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>