server: Make sure view names are always schema-qualified
The way view creation works is as follows:

1. When views are created, we schema-qualify the name if necessary and
   then put the view in a map of `uncompiled_views` in `SqlIncorporator`
2. When a CREATE CACHE fails due to a table not found error, we check to
   see if that table not found error schema-resolves to an uncompiled
   view we know about. If it does, we remove that view from the map of
   uncompiled views, compile it, store it in the map of known relations,
   and then run the CREATE CACHE again

There's a sneaky issue in step 2, however - if the reference to the view
in the `CREATE CACHE` is not schema qualified, then the table not found
error actually comes from the schema resolution rewrite pass, and
hence *doesn't contain a schema* - we'd correctly resolve the view
reference, but then use the *non-schema-qualified* name when storing the
view in the map of known relations! Then, if we got a DROP VIEW for that
view, it'd fail to resolve due to the view being stored using a
non-schema-qualified name.

Instead, this commit stores the (properly schema-qualified) `name` as an
extra field on `UncompiledView`, and uses *that* as the name we store
the view under when compiling it.

Fixes: ENG-2673
Change-Id: Ie26668d80a69807c031ee6104888a32fe422ce52
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4487
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>