ddl_vertical: Skip steps after the failing step
This optimizes shrinking by skipping operations that occur after the
operation that triggered the test failure. This was a bit tricky to do,
because the proptest framework doesn't offer any convenient ways to
share state between the test function and the `Strategy`/`ValueTree`
structs that are used for test case generation and shrinking. I was able
to make it work without too much ugliness via interior mutability.

Refs: ENG-2613
Change-Id: I3b2695fed22d897263c5945328a51cd84a4908d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4492
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>