psql-srv: Implement SCRAM-SHA-256 authentication
This commit implements SCRAM-SHA-256 authentication in psql-srv. This is
mostly a direct implementation of the algorithm described in RFC5802[0],
using the hashing mechanism specified in RFC7677[1]. I elected to NIH it
despite the existence of the scram crate[2] because that crate seems
poorly-maintained and does not support channel-binding, which I would
like to add support for next (though this commit does not implement
channel binding, there's a couple of TODO comments referencing
ENG-2698).

[0]: https://www.rfc-editor.org/rfc/rfc5802
[1]: https://www.rfc-editor.org/rfc/rfc7677
[2]: https://crates.io/crates/scram

Fixes: ENG-2683
Change-Id: Ibd2fb478c05eef8c44cc3cf492706b4c9a3721c8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4504
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>