domains: Initialize persistent state in a different thread.
Currently, when we handle the `DomainRequest::Ready` message, we
materialize the state of the given node as part of the process of
handling the request.
The problem with that approach is that we've found that for very large
tables and during recovery, bootstrapping the persistent state takes a
*really* long time, which ends up causing the RPC call of that requet to
 time out.

To work around that, we are now moving the initialization of the
persistent state to a separate thread, so that we can return
immediately.
In doing so, we defer marking the node as `ready` (so the node index is
kept in the `not_ready` set in the `Domain`), which we then use to
ignore any `DomainRequest::StartReplay` message we might get.

If the base node gets removed while the persistent state is being
initialized, the initialization won't be interrupted and the persistent
state will be torn down right after initialization.
We do it this way because:
1. Interrupting the initialization of the persistent state is hard to
get right, as we would have to clean up the database files that might
remain.
2. This is the current behavior: we can't process a request to remove a
base node until we process the request to initialize said node. So, no
real change there.

As a side note:
 1. `Base::unique_keys` now directly returns
`Vec<Box<[usize]>>`, since it was only being called during
`DomainRequest::Ready`, so it was better to return what we already need
directly.
2. There are 2 clustertests that fail in CI but pass locally, which have
 been fixed by duplicating the time they waited for workers to be
 healthy (from 60 secs to 120 secs).

Fixes: ENG-2352
Change-Id: I2923b15e6012adeed606d51c8cf23eb54fb59667
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4514
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>