readyset-psql: Allow configuring authentication method
Allow configuring the authentication method to use for postgresql
clients in readyset-psql, and default to the newly-supported
SCRAM-SHA-256 authentication method since it's got better security.
This piggy-backs on the new `readyset::psql::Options` struct which was
previously TLS-only, since it provides a nice mechanism for
psql-specific options.

Release-Note-Core: ReadySet running in postgresql mode now defaults to
  SCRAM-SHA-256 authentication for all clients. To force ReadySet to use
  cleartext authentication (the previous behavior) instead, pass
  --postgres-authentication-method=cleartext or set the
  `POSTGRES_AUTHENTICATION_METHOD` environment variable to `cleartext`
  when running the `readyset` binary
Fixes: ENG-2699
Change-Id: Ifa9c6b3bac52972602a66455f564e894edb9c054
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4505
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>