ddl_vertical: Optimize shrinking by avoiding duplicate test cases
Prior to this change, we would often unnecessarily re-test the same test
case more than once during the shrinking process. The issue is that the
proptest framework treats `simplify` and `complicate` more like a binary
search, so the framework re-runs the test again after calling
`complicate`. But in the `ValueTree` impl prior to this change, our
`complicate` implementation simply undid the last call to `simplify`,
which would cause us to re-run the same test case that we already ran
prior to the last call to `simplify`.

To fix this, I've factored out the code that tries removing another op,
and we now call that code from both `simplify` and `complicate`. The
`complicate` code still adds back the last op that was removed, since
the fact that `complicate` is being called implies that the last op was
necessary to reproduce the failure; however, `complicate` then goes
ahead and tries to find another op that we can shrink out, thus skipping
the unnecessary intermediate step where we re-run the previously tested
case.

In a real-world example I tried, this optimization reduces the time
needed to find a minimal failing case by about half.

Refs: ENG-2613
Change-Id: I27612756354855a513af6ef5aed75019a668d6b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4508
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI