psql-srv: Support SCRAM-SHA-256-PLUS with channel binding
Implement support for SCRAM-SHA-256-PLUS authentication with
tls-server-end-point channel binding in psql-srv, based on section 6 of
RFC5802[0]. The gist here is that we use the
"tls-server-end-point" (provided by the `tls_server_end_point()` method
on the TlsStream for the client) as part of the overall input to the
HMAC for the SCRAM handshake, and validate that that's been done as well
by the client (by checking the channel binding data against the
`cbind-input` field of the `client-final-message`).

Note that even though the RFC says that servers MUST implement the
`tls-unique` channel binding type, we don't do that here because
PostgreSQL *always* uses tls-server-end-point.

[0]: https://www.rfc-editor.org/rfc/rfc5802#section-6

Fixes: ENG-2698
Change-Id: I1e03e275e028722f9aa56a534b60b39691dd2d03
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4513
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>