ddl_vertical: Fix missing state update for CreateSimpleView
The lack of this state update has not actually caused any test failures
I've seen, presumably because the odds of reusing table/view names twice
in the same test case is very low right now. Still worth fixing though,
and may become more relevant in the future if we change the test in the
future to encourage more reuse of names.

Change-Id: Ic452b13e95dc1336a48dc82642a762892dcbdd7b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4523
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>