ddl_vertical: Add support for generating and creating join views
This completes support for the new "create join view" operation.
Implementation was all pretty straightforward, and this seems to work
fine as-is; haven't found any new bugs as a result of adding this
operation yet. I'll leave it running test cases for a while this evening
though, in case it finds anything interesting given more time.

Refs: ENG-2320
Change-Id: I01ed3406c2a13b953cad1446bbb9f39afe9126d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4524
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>