treewide: Remove LogContext and readyset-tracing log wrapper
The LogContext support added to readyset-tracing, along with the
treewide change to import tracing macros via the readyset-tracing crate
rather than via tracing, was done as a way to log the name of the
deployment as an extra field in every single log message. After running
with this for a while, it feels like this mostly just adds a ton of
noise to log messages during local development, plus additional
maintenance and review burden of having to import readyset_tracing
instead of tracing everywhere - when we can just add the deployment name
via the log *ingestion* method as part of the deployment infrastructure,
rather than doing it in-application (or, if we really want to do it in
the application, just wrap the `main` function in a span that logs the
name of the deployment).

This commit removes the log-wrapper abstraction from readyset-tracing,
and replaces the import of readyset-tracing with an import of tracing
treewide (that second bit is essentially a revert of commit
514ca30bc67b3437c061f1bac1f708b06a641cc8).

Refs: ENG-2703#	public/replicators/tests/tests.rs
Change-Id: I4b71f35e0743afdc1742d6cd1dd5c99e4edfdc74
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4536
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>