sql: Make uncompiled views participate in invalidation
Today, if the compilation of a query "passes over" a schema when trying
to resolve a reference to a table, we store that hypothetical table in a
list of tables that, if they are ever *created*, should invalidate that
query - for example, if we have schema_search_path = [s1,s2] and `s2.t`
exists but not `s1.t` and we create a query like `SELECT * FROM t`, we
will store `s1.t` in the list of invalidating tables for that query, so
that the query is invalidated if `s1.t` is ever created. This works fine
for *tables*, but because views are created via a different
mechanism (they go through `uncompiled_views`, to allow us to avoid
creating nodes for views before they're actually referenced by a cache),
view references didn't benefit from the same invalidating logic. This
commit modifies view compilation to properly participate in table
invalidation, by:

- Making *references* to uncompiled views in queries properly
  participate in invalidation, which requires passing a separate list of
  known-existent uncompiled views to `rewrite` since they can't go in
  `view_schemas`, since we don't know what their fields will be (eg if
  they have a `*`) until we actually compile them
- Making view creation consult `queries_to_invalidate_for_table` to drop
  queries that should've resolved to that view
- Doing some juggling on lazy-view-compilation to make sure that
  we *always* get relations we could've referenced by the query if they
  existed inserted into `invalidating_tables`.

Fixes: ENG-2703
Change-Id: I265ff7cae39f87a9b86056028fe3378edf469c53
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4538
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>