Adapter: Fix reorder_params() in make_keys()
Changes the logic of reorder_params() so that the indices recorded in
the reorder_numbered_placeholders() pass are used to index into the
given parameters rather than the reordered parameter vector.

The new logic also supports parameter reuse in the rewrite pass (e.g.,
'SELECT * FROM t WHERE a = $1 AND b = $1'). However, parameter reuse is
not yet supported across the adapter.

Change-Id: I0f9ab7ce43d6a665f9d48fbd92edafe5d8aa703e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4556
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>