Adapter: Update all placeholders in NumberPlaceholdersVisitor
The current NumberPlaceholdersVisitor attempted to maintain existing
DollarNumber placeholder values. This isn't necessary and is also
problematic when the WHERE IN rewrite pass creates a gap in the final
set of DollarNumber placeholders.

The NumberPlaceholdersVisitor is updated so that the final set of
placeholders is contiguous. We do not need to store placeholder value
remappings as the values are already ordered.

Adds a logictest for this issue, which is tested under the mysql
logictests because of Eng-2962. readyset-mysql will still parse
positional params.

Change-Id: I583143754fbc50ffdd68d5c15078cb56adc4e6a1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4557
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>