dataflow-expression: Replace negated binops with a NOT node
As we move towards doing a little bit more work during
lowering (specifically, deciding when and how to coerce values), it's
going to be a lot easier to deal with stuff if the target language is
simpler. To that end, this commit removes all the "negated" versions of
the (lowered) binary operators in dataflow-expression, and introduces a
new unary "Not" expr node, which we wrap around the binary operator node
when the surface AST operator has been negated.

This actually revealed a *bug* in the way we used to handle NOT, where
`NOT -1` would incorrectly evaluate to true in mysql when it should've
evaluated to false. The test that (incorrectly) asserted the wrong
behavior here has been fixed.

Change-Id: Iad313049a7e3b052169a3c41de7f855e348d4611
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4560
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>