dataflow-expression: Introduce a non-generative pg oracle test
Add a new test suite to dataflow-expression that runs through a
hard-coded list of expressions, evaluates each expression using
dataflow-expression and postgres, and compares the results. This is
ideally a slightly lower-effort way of testing the evaluation of a *lot*
of different expressions, which'll be important as we refactor the
internals of dataflow-expression.

Change-Id: I9ce028f19cf9c3444f067f00bb586c576e1ca967
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4561
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>