ddl_vertical: Add DropEnum operation
This adds the ability to drop ENUM types after they've been created. Now
that I've created this, I'm actually not totally sure how valuable this
is on its own since we don't have any straightforward way to check that
the type has actually been dropped, but for now the code works and the
tests continue to pass so I will think about whether there's a nice way
to make this more useful in future commits.

Refs: ENG-2710
Change-Id: I849a81e1a2721c544edee3005023d63c1f078218
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4562
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>