ddl_vertical: Factor out tables_using_type function
Right now this refactor is not a big win on its own, but this new
function will be reused in upcoming commits so that we can find which
caches we need to explicitly recreate after altering enum types. In the
meantime it also simplifies the precondition function a little bit.

Change-Id: I11ac8e20ad067589db9ca2a99b3f500b2f871d26
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4576
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>