data: Always allow coercing to unknown
Especially as we start doing coercions explicitly in more places in
dataflow expression, it makes the most sense to be permissive rather
than restrictive here and make "coerce to unknown" just a no-op rather
than returning an error (which it still does in the individual coerce
modules as we still have to pattern match on it somehow). This is
especially relevant for arrays, since in practice it's useful to be able
to have an operation that looks like "coerce this (maybe string-literal)
value to an array, but I don't want to know or care what the *member*
type of that array is".

Change-Id: If0bc4b4a886d7719d328af3e5d4100ac77628be9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4572
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>