dataflow-expression: Decide coercion for binops at lowering
We'd like to refactor dataflow-expression such that all decisions about
how and when to coerce values for expressions happens at lowering time,
via emitting a cast node in the lowered expression tree, rather than
having to make those decisions at eval time. This is especially
important for things with nested types like arrays, where the decision
about how to perform coercion is non-trivial.

This commit starts us down that path, by emitting coerce nodes for the
lhs and rhs of all binary operators at lowering time, and removing
coercion for both from expression evaluation. This also includes a
smattering of extra oracle tests for both mysql and postgresql, the
latter including some extra support code to create custom enums, since
enums are especially interesting in the context of expressions. In
general, the idea here is that this is trying as hard as possible to
just replicate existing coercion logic rather than fix any bugs where
it's known to be incorrect - there are a couple of expressions I've
added to the mysql oracle test that document some issues in our
expression eval that I found during this process.

Fixes: ENG-2762
Change-Id: Iff6f5051859d6254153ecc3d3a1516f858eeb0da
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4573
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>