Unsupported Placeholders: Detect unsupported mixed comparisons
Adds detection of placeholders in unsupported mixed comparisons to the
detect_unsupported_placeholders() sql pass. This change only applies if
experimental mixed comparison support is disabled. If it is enabled,
this CL does not change the result of the sql pass.

If we have a query with a mix of equality and ordering comparisons, we
could consider either the placeholders in the equality or the ordering
comparisons as unsupported. In this CL, we always flag the placeholders
in the ordering comparisons as unsupported. This will affect inlining
that is done automatically in the future, but we will still be able to
choose which sets of parameters to inline if the inlining is done
manually.

This CL enables manual inlined of queries with mixed comparisons because
we do not look to see if an inlined version of a query exists unless it
fails the detect_unsupported_placeholders() pass.

Change-Id: I86ef9fd75c9c8304c703d76ee92be8263a0625dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4574
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>