ddl_vertical: Add support for appending new values to ENUM types
This adds the new AddEnumValue operation, which adds a value to an ENUM
(without specifying the position, so the new value gets added to the
end).

This seems to trigger a bug somewhere, but I haven't quite figured out
where it's happening yet, so it could be a test bug or a ReadySet bug.

Refs: ENG-2711
Change-Id: I8a8bfbfc477a17391e8bf5c4bbe71c165d5a43f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4577
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>