ddl_vertical: Reset next_step_idx to 0 for each test case
I had previously only tested this in cases where a known failure from
the proptest-regressions file triggers a failed test immediately. I
neglected to consider that the Cell is shared across multiple test cases
and keeps being incremented, which can lead to a panic when we try to
update the shrinking state for very high indexes outside the bounds of
`currently_included`.

Change-Id: Ibf2608b30dfb8f12ea210b350a96eb773bf1b474
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4578
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>