ddl_vertical: Temporarily disable AddEnumValue operation
This operation sometimes triggers ENG-2763, so I'm temporarily disabling
it for now, so as to keep the test passing (though, it's not running in
CI yet, so this is not a huge deal, but it'll enable me to keep adding
other operations in the meantime without running into the same failure).

Once ENG-2763 is fixed, we should revert this commit.

Refs: ENG-2763
Change-Id: Icc09dda5f6d39a1c35beebf2a3dadd88f7eee116
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4580
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>